Split off main site comment controller from discussion controller
