+++ plugins/ShinyCMS/app/views/shinycms/comments/new.html.erb
@@ -1,0 +1,2 @@
+<%= render partial: 'discussions/comment/form',
+           locals: { discussion: @discussion } %>
