+++ plugins/ShinyCMS/app/views/shinycms/comments/show.html.erb
@@ -1,0 +1,2 @@
+<%= render partial: 'discussions/comment',
+           locals: { discussion: @discussion, comment: @comment, depth: 0 } %>
